Issue 26 completed
Issues 26 is in the checkout block when we give the long size emails then the emails is not fitting into it properly. So i changed width size to 100% but actually i tried to find a flexible where the block will be changing width size based on length of email but it isn't working out so i am going with some simple solution.